XAI/requirements.txt
# ==============================================================================
# XAI-FER Project Requirements
# ==============================================================================
# Install with: pip install -r requirements.txt
# ==============================================================================

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Models
transformers>=4.30.0
timm>=0.9.0

# Image Processing
Pillow>=9.0.0

# Visualization
matplotlib>=3.7.0
tqdm>=4.60.0

# XAI Methods — CNN (CAM-based)
grad-cam>=1.4.0

# XAI Methods — Agnostic (LIME/SHAP)
lime>=0.2.0
shap>=0.42.0

# Required by LIME
scikit-image>=0.20.0

# Required by SHAP
scikit-learn>=1.3.0
scipy>=1.10.0